move redteam commands into redteam folder
