+++ src/onboarding.ts
@@ -6,9 +6,9 @@ import chalk from 'chalk';
 import dedent from 'dedent';
 import fs from 'fs';
 import path from 'path';
-import { redteamInit } from './commands/redteam/init';
 import { getEnvString } from './envars';
 import logger from './logger';
+import { redteamInit } from './redteam/commands/init';
 import telemetry, { type EventValue } from './telemetry';
 import type { EnvOverrides } from './types';
 import { getNunjucksEngine } from './util/templates';
