+++ src/redteam/commands/generate.ts
@@ -6,19 +6,11 @@ import yaml from 'js-yaml';
 import path from 'path';
 import invariant from 'tiny-invariant';
 import { z } from 'zod';
+import { synthesize } from '..';
 import { disableCache } from '../../cache';
 import cliState from '../../cliState';
 import { resolveConfigs } from '../../config';
 import logger from '../../logger';
-import { synthesize } from '../../redteam';
-import {
-  REDTEAM_MODEL,
-  DEFAULT_PLUGINS as REDTEAM_DEFAULT_PLUGINS,
-  ADDITIONAL_PLUGINS as REDTEAM_ADDITIONAL_PLUGINS,
-  DEFAULT_STRATEGIES,
-  ADDITIONAL_STRATEGIES,
-} from '../../redteam/constants';
-import { shouldGenerateRemote } from '../../redteam/util';
 import telemetry from '../../telemetry';
 import type { TestSuite, UnifiedConfig } from '../../types';
 import type { RedteamStrategyObject, SynthesizeOptions } from '../../types/redteam';
@@ -26,6 +18,14 @@ import type { RedteamFileConfig, RedteamCliGenerateOptions } from '../../types/r
 import { printBorder, setupEnv } from '../../util';
 import { writePromptfooConfig } from '../../util/config';
 import { RedteamGenerateOptionsSchema, RedteamConfigSchema } from '../../validators/redteam';
+import {
+  REDTEAM_MODEL,
+  DEFAULT_PLUGINS as REDTEAM_DEFAULT_PLUGINS,
+  ADDITIONAL_PLUGINS as REDTEAM_ADDITIONAL_PLUGINS,
+  DEFAULT_STRATEGIES,
+  ADDITIONAL_STRATEGIES,
+} from '../constants';
+import { shouldGenerateRemote } from '../util';
 
 export async function doGenerateRedteam(options: RedteamCliGenerateOptions) {
   setupEnv(options.envFile);
