+++ src/redteam/commands/plugins.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import type { Command } from 'commander';
 import logger from '../../logger';
-import { ALL_PLUGINS, DEFAULT_PLUGINS, subCategoryDescriptions } from '../../redteam/constants';
+import { ALL_PLUGINS, DEFAULT_PLUGINS, subCategoryDescriptions } from '../constants';
 
 export function pluginsCommand(program: Command) {
   program
