+++ test/redteam/commands/generate.test.ts
@@ -1,8 +1,8 @@
 import fs from 'fs';
-import { doGenerateRedteam } from '../../../src/commands/redteam/generate';
 import * as configModule from '../../../src/config';
 import logger from '../../../src/logger';
 import { synthesize } from '../../../src/redteam';
+import { doGenerateRedteam } from '../../../src/redteam/commands/generate';
 import type { RedteamCliGenerateOptions } from '../../../src/types/redteam';
 import { writePromptfooConfig } from '../../../src/util/config';
